multithreading/src/test/java/concurrent/SimpleBlockingQueueTest.java
package concurrent;

import static org.junit.Assert.*;

import org.junit.Test;

import java.util.ArrayList;
import java.util.Arrays;
import java.util.List;
import java.util.concurrent.CopyOnWriteArrayList;

public class SimpleBlockingQueueTest {
    @Test
    public void test2() throws InterruptedException {
        final CopyOnWriteArrayList<Integer> buffer = new CopyOnWriteArrayList<>();
        final SimpleBlockingQueue<Integer> queue = new SimpleBlockingQueue<>();
        Thread producer = new Thread(
                () -> {
                    try {
                        for (int i = 0; i < 5; i++) {
                            queue.offer(i);
                        }
                    } catch (InterruptedException e) {
                        Thread.currentThread().interrupt();
                    }
                }
        );
        producer.start();
        Thread consumer = new Thread(
                () -> {
                    while (!queue.isEmpty() || !Thread.currentThread().isInterrupted()) {
                        try {
                            buffer.add(queue.poll());
                        } catch (InterruptedException e) {
                            Thread.currentThread().interrupt();
                        }
                    }
                }
        );
        consumer.start();
        producer.join();
        consumer.interrupt();
        consumer.join();
        assertEquals(Arrays.asList(0, 1, 2, 3, 4), buffer);
    }

    @Test
    public void test() throws InterruptedException {
        final SimpleBlockingQueue queue = new SimpleBlockingQueue<Integer>();
        final List<Integer> result = new ArrayList<>();
        Thread thread1 = new Thread(() -> {
            try {
                result.add((Integer) queue.poll());
            } catch (InterruptedException e) {
                e.printStackTrace();
            }
        });
        result.add(1);
        Thread thread2 = new Thread(() -> {
            try {
                queue.offer(2);
            } catch (InterruptedException e) {
                e.printStackTrace();
            }
        });
        thread1.start();
        thread2.start();
        thread1.join();
        thread2.join();
        assertEquals(List.of(1, 2), result);
    }
}
